requirements.txt
# ---------------------
# Core runtime 
# ---------------------
httpx>=0.27,<0.28
tenacity>=8.2,<9
pydantic>=2.7,<3
click>=8.1,<9
tqdm>=4.66,<5
numpy>=1.26,<3
shapely>=2.0,<3
pyproj>=3.6,<4
orjson>=3.10,<4            # fast JSON, kinda helpful
python-dateutil>=2.9,<3

# ---------------------
# Storage: S3 now, GCS/Azure later 
# ---------------------
boto3>=1.34,<2
botocore>=1.34,<2
s3transfer>=0.10,<0.11
fsspec>=2024.6.1
s3fs>=2024.6.0
python-dotenv>=1.0,<2

# ---------------------
# Logging / Telemetry 
# ---------------------
rich>=13.7,<14
structlog>=24.1,<25

# ---------------------
# Concurrency / performance
# ---------------------
anyio>=4.4,<5

# --------------------- 
# Testing & Dev
# ---------------------
pytest>=8,<9
pytest-cov>=5,<6
mypy>=1.10,<2
ruff>=0.5,<1
types-python-dateutil>=2.9.0.20240316
types-requests>=2.32.0.20240712

# ---------------------
# CLI UX
# ---------------------
typer>=0.12,<0.13

# ---------------------
# Heavy Geo (commented until needed; requires system GDAL on macOS - will install that later
# ---------------------
rasterio>=1.3.10,<1.4
rio-cogeo>=4.1,<5
pyogrio>=0.9,<1
geopandas>=0.14,<1
fiona>=1.9,<2

